Combine test_rms and test_rms_config
